requirements.txt
gradio>=4.0.0
numpy>=1.20.0
pandas>=1.3.0
torch>=1.10.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
requests>=2.26.0
tqdm>=4.62.0
scikit-learn>=1.0.0
PyYAML>=6.0